refactor(automation): move automation logic to internal/core layer (#14)
