+++ internal/core/automation/automation.go
@@ -8,25 +8,6 @@ import (
 	"github.com/mdgspace/sysreplicate/internal/domain"
 )
 
-type AutomationData struct {
-	SystemDServices []SystemDUnit `json:"systemd_services"`
-	SystemDTimers   []SystemDUnit `json:"systemd_timers"`
-	UserCronjobs    []Cronjob     `json:"user_cronjobs"`
-	SystemCronjobs  []Cronjob     `json:"system_cronjobs"`
-}
-type SystemDUnit struct {
-	Name        string `json:"name"`
-	Path        string `json:"path"`
-	Content     string `json:"content"`
-	UnitType    string `json:"unit_type"` ////saare service, timer and target available ere
-	IsEnabled   bool   `json:"is_enabled"`
-	IsActive    bool   `json:"is_active"`
-}
-type Cronjob struct {
-	Path    string `json:"path"`
-	Content string `json:"content"`
-	Type    string `json:"type"` //.//user, system, cron_d
-}
 type AutomationManager struct {
 	username string
 }
@@ -40,14 +21,14 @@ func NewAutomationManager() *AutomationManager {
 		username: username,
 	}
 }
-func (am *AutomationManager) DetectAutomation() (*AutomationData, error) {
+func (am *AutomationManager) DetectAutomation() (*domain.AutomationData, error) {
 	fmt.Println("Detecting automation files...")
 	
-	data := &AutomationData{
-		SystemDServices: make([]SystemDUnit, 0),
-		SystemDTimers:   make([]SystemDUnit, 0),
-		UserCronjobs:    make([]Cronjob, 0),
-		SystemCronjobs:  make([]Cronjob, 0),
+	data := &domain.AutomationData{
+		SystemDServices: make([]domain.SystemDUnit, 0),
+		SystemDTimers:   make([]domain.SystemDUnit, 0),
+		UserCronjobs:    make([]domain.Cronjob, 0),
+		SystemCronjobs:  make([]domain.Cronjob, 0),
 	}
 	
 	systemdServices, systemdTimers, err := am.detectSystemDUnits()
